requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
DAWG-Python==0.7.2
docopt==0.6.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
multidict==6.1.0
propcache==0.2.0
pyaes==1.6.1
pyasn1==0.6.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==21.7
requests==2.32.3
rsa==4.9
setuptools==75.5.0
sniffio==1.3.1
telegram==0.0.1
Telethon==1.37.0
urllib3==2.2.3
watchdog==6.0.0
yarl==1.18.0
